Use our own json handle for decoding (#765)
* Use our own json handle for decoding.

* Have encoding, decoding fuctions in the same file.
